Overhaul of CONTRIBUTING document
I took a lot of liberties with this because it seemed way too long for such a
small project, and we're not likely to enforce the finer points, anyway. The
more hoops someone has to jump through to contribute, the less likely they are
to do so.

So, basically, I removed a lot of things and added none.

I am open to discussing this change if it proves contentious.